LPS-100562 Namespace sidebar components
